\name{binary_cre}
\alias{binary_cre}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Internal function for binary model.
}
\description{
An internal function.
}
\usage{
binary_cre(Z_mat = Z_mat, first.order = first.order, 
          home.field, control = control)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{Z_mat}{
data frame.
}
  \item{first.order}{
logical
}
  \item{home.field}{
logical
}
  \item{control}{
list
}
}
