\name{N_mov}
\alias{N_mov}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Internal Function for Normal MOV model
}
\description{
Internal Function for Normal MOV model
}
\usage{
N_mov(Z_mat = Z_mat, first.order = TRUE, 
   home.field = home.field, control = control)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{Z_mat}{
data frame
}
  \item{first.order}{
logical
}
  \item{home.field}{
logical
}
  \item{control}{
list
}
}
