\name{NB_cre}
\alias{NB_cre}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Internal Function for Normal-Binary Model
}
\description{
Internal Function for Normal-Binary Model
}
\usage{
NB_cre(Z_mat = Z_mat, first.order = first.order, 
   home.field = home.field, control = control)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{Z_mat}{
data frame
}
  \item{first.order}{
logical
}
  \item{home.field}{
logical
}
  \item{control}{
list
}
}
