\name{PB_cre}
\alias{PB_cre}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Internal Function for Poisson-binary Model
}
\description{
Internal Function for Poisson-binary Model
}
\usage{
PB_cre(Z_mat = Z_mat, first.order = first.order,
   home.field = home.field, control = control, 
   game.effect = game.effect)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{Z_mat}{
data frame
}
  \item{first.order}{
logical
}
  \item{home.field}{
logical
}
  \item{control}{
list
}
  \item{game.effect}{
logical
}
}
