\name{poisson_cre}
\alias{poisson_cre}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Internal Function for Poisson Model
}
\description{
Internal Function for Poisson Model
}
\usage{
poisson_cre(Z_mat = Z_mat, first.order = first.order, 
   control = control, game.effect = game.effect, 
   home.field = home.field)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{Z_mat}{
data frame
}
  \item{first.order}{
logical
}
  \item{control}{
logical
}
  \item{game.effect}{
logical
}
  \item{home.field}{
logical
}
}
